requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
jsonschema==4.9.0
lxml==4.9.1
nordvpn-switcher==0.3.0
numpy==1.23.2
openpyxl==3.0.10
outcome==1.2.0
pandas==1.4.4
pathlib==1.0.1
psutil==5.9.1
pycparser==2.21
pycutils==0.0.7
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
random-user-agent==1.0.1
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
validator-collection==1.5.0
webdriver-manager==3.8.3
wsproto==1.1.0
